+++ git-worktree/tests/worktree/fs/cache/mod.rs
@@ -1,0 +1,4 @@
+mod create_directory;
+
+#[allow(unused)]
+mod ignore_and_attributes;
